requirements.txt
backports.statistics==0.1.0
Cython==0.29.28
dl==0.1.0
gpiozero==1.6.2
ipaddr==2.2.0
luma==2.1.3
luma.oled==3.8.1
#luma.core
#RPi.GPIO
mysql_connector_repackaged==0.3.1
numpy==1.16.2
ordereddict==1.1
wincertstore==0.2.1
